Add Mod and refactor to smaller files
