+++ src/common/names.ts
@@ -1,544 +1,15 @@
+import { CAT_NAMES } from '../panel/pets/cat';
+import { CLIPPY_NAMES } from '../panel/pets/clippy';
+import { COCKATIEL_NAMES } from '../panel/pets/cockatiel';
+import { CRAB_NAMES } from '../panel/pets/crab';
+import { DOG_NAMES } from '../panel/pets/dog';
+import { MOD_NAMES } from '../panel/pets/mod';
+import { ROCKY_NAMES } from '../panel/pets/rocky';
+import { DUCK_NAMES } from '../panel/pets/rubberduck';
+import { SNAKE_NAMES } from '../panel/pets/snake';
+import { TOTORO_NAMES } from '../panel/pets/totoro';
+import { ZAPPY_NAMES } from '../panel/pets/zappy';
 import { PetType } from './types';
-
-export const CAT_NAMES: ReadonlyArray<string> = [
-    'Bella',
-    'Charlie',
-    'Molly',
-    'Coco',
-    'Ruby',
-    'Oscar',
-    'Lucy',
-    'Bailey',
-    'Milo',
-    'Daisy',
-    'Archie',
-    'Ollie',
-    'Rosie',
-    'Lola',
-    'Frankie',
-    'Roxy',
-    'Poppy',
-    'Luna',
-    'Jack',
-    'Millie',
-    'Teddy',
-    'Cooper',
-    'Bear',
-    'Rocky',
-    'Alfie',
-    'Hugo',
-    'Bonnie',
-    'Pepper',
-    'Lily',
-    'Tilly',
-    'Leo',
-    'Maggie',
-    'George',
-    'Mia',
-    'Marley',
-    'Harley',
-    'Chloe',
-    'Lulu',
-    'Missy',
-    'Jasper',
-    'Billy',
-    'Nala',
-    'Monty',
-    'Ziggy',
-    'Winston',
-    'Zeus',
-    'Zoe',
-    'Stella',
-    'Sasha',
-    'Rusty',
-    'Gus',
-    'Baxter',
-    'Dexter',
-    'Willow',
-    'Barney',
-    'Bruno',
-    'Penny',
-    'Honey',
-    'Milly',
-    'Murphy',
-    'Simba',
-    'Holly',
-    'Benji',
-    'Henry',
-    'Lilly',
-    'Pippa',
-    'Shadow',
-    'Sam',
-    'Lucky',
-    'Ellie',
-    'Duke',
-    'Jessie',
-    'Cookie',
-    'Harvey',
-    'Bruce',
-    'Jax',
-    'Rex',
-    'Louie',
-    'Jet',
-    'Banjo',
-    'Beau',
-    'Ella',
-    'Ralph',
-    'Loki',
-    'Lexi',
-    'Chester',
-    'Sophie',
-    'Chilli',
-    'Billie',
-    'Louis',
-    'Scout',
-    'Cleo',
-    'Purfect',
-    'Spot',
-    'Bolt',
-    'Julia',
-    'Ginger',
-    'Daisy',
-    'Amelia',
-    'Oliver',
-    'Ghost',
-    'Midnight',
-    'Pumpkin',
-    'Shadow',
-    'Binx',
-    'Riley',
-    'Lenny',
-    'Mango',
-    'Alex',
-    'Boo',
-    'Botas',
-    'Romeo',
-    'Bob',
-    'Clyde',
-    'Simon',
-    'Mimmo',
-    'Carlotta',
-    'Felix',
-    'Duchess',
-];
-
-export const DOG_NAMES: ReadonlyArray<string> = [
-    'Bella',
-    'Charlie',
-    'Max',
-    'Molly',
-    'Coco',
-    'Buddy',
-    'Ruby',
-    'Oscar',
-    'Lucy',
-    'Bailey',
-    'Milo',
-    'Daisy',
-    'Archie',
-    'Ollie',
-    'Rosie',
-    'Lola',
-    'Frankie',
-    'Toby',
-    'Roxy',
-    'Poppy',
-    'Luna',
-    'Jack',
-    'Millie',
-    'Teddy',
-    'Harry',
-    'Cooper',
-    'Bear',
-    'Rocky',
-    'Alfie',
-    'Hugo',
-    'Bonnie',
-    'Pepper',
-    'Lily',
-    'Leo',
-    'Maggie',
-    'George',
-    'Mia',
-    'Marley',
-    'Harley',
-    'Chloe',
-    'Lulu',
-    'Jasper',
-    'Billy',
-    'Nala',
-    'Monty',
-    'Ziggy',
-    'Winston',
-    'Zeus',
-    'Zoe',
-    'Stella',
-    'Sasha',
-    'Rusty',
-    'Gus',
-    'Baxter',
-    'Dexter',
-    'Diesel',
-    'Willow',
-    'Barney',
-    'Bruno',
-    'Penny',
-    'Honey',
-    'Milly',
-    'Murphy',
-    'Holly',
-    'Benji',
-    'Henry',
-    'Lilly',
-    'Pippa',
-    'Shadow',
-    'Sam',
-    'Buster',
-    'Lucky',
-    'Ellie',
-    'Duke',
-    'Jessie',
-    'Cookie',
-    'Harvey',
-    'Bruce',
-    'Jax',
-    'Rex',
-    'Louie',
-    'Bentley',
-    'Jet',
-    'Banjo',
-    'Beau',
-    'Ella',
-    'Ralph',
-    'Loki',
-    'Lexi',
-    'Chester',
-    'Sophie',
-    'Billie',
-    'Louis',
-    'Charlie',
-    'Cleo',
-    'Spot',
-    'Harry',
-    'Bolt',
-    'Ein',
-    'Maddy',
-    'Ghost',
-    'Midnight',
-    'Pumpkin',
-    'Shadow',
-    'Sparky',
-    'Linus',
-    'Cody',
-    'Slinky',
-    'Toto',
-    'Balto',
-    'Golfo',
-    'Pongo',
-    'Beethoven',
-    'Hachiko',
-    'Scooby',
-    'Clifford',
-    'Astro',
-    'Goofy',
-    'Chip',
-    'Einstein',
-    'Fang',
-    'Truman',
-    'Uggie',
-    'Bingo',
-    'Blue',
-    'Cometa',
-    'Krypto',
-    'Huesos',
-    'Odie',
-    'Snoopy',
-    'Aisha',
-    'Moly',
-    'Chiquita',
-    'Chavela',
-    'Tramp',
-    'Lady',
-    'Puddles',
-];
-
-export const CRAB_NAMES: ReadonlyArray<string> = [
-    'Ferris',
-    'Pinchy',
-    'Grabby',
-    'Big Red',
-    'Crabby',
-    'Buddy',
-    'Ruby Red',
-    'Oscar',
-    'Lucy',
-    'Bailey',
-    'Crabito',
-    'Percy',
-    'Rocky',
-    'Mr. Krabs',
-    'Shelly',
-    'Santa Claws',
-    'Clawdia',
-    'Scuttle',
-    'Snappy',
-    'Hermit',
-    'Horseshoe',
-    'Snapper',
-    'Coconut',
-    'Sebastian',
-    'Abby',
-    'Bubbles',
-    'Bait',
-    'Big Mac',
-    'Biggie',
-    'Claws',
-    'Copper',
-    'Crabette',
-    'Crabina',
-    'Crabmister',
-    'Crusty',
-    'Crabcake',
-    'Digger',
-    'Nipper',
-    'Pincer',
-    'Poopsie',
-    'Recluse',
-    'Salty',
-    'Squirt',
-    'Groucho',
-    'Grumpy',
-    'Lenny Krabitz',
-    'Leonardo DaPinchy',
-    'Peeves',
-    'Penny Pincher',
-    'Prickl',
-];
-
-export const CLIPPY_NAMES: ReadonlyArray<string> = [
-    'Clippy',
-    'Karl Klammer',
-    'Clippy Jr.',
-    'Molly',
-    'Coco',
-    'Buddy',
-    'Ruby',
-    'Oscar',
-    'Lucy',
-    'Bailey',
-];
-
-export const TOTORO_NAMES: ReadonlyArray<string> = [
-    'Totoro',
-    'トトロ',
-    'Max',
-    'Molly',
-    'Coco',
-    'Buddy',
-    'Ruby',
-    'Oscar',
-    'Lucy',
-    'Bailey',
-    'Big fella',
-];
-
-export const SNAKE_NAMES: ReadonlyArray<string> = [
-    'Sneaky',
-    'Mr Slippery',
-    'Hissy Elliott',
-    'Molly',
-    'Coco',
-    'Buddy',
-    'Ruby',
-    'Bailey',
-    'Max',
-    'Seb',
-    'Kaa',
-    'Mr Hiss',
-    'Miss Hiss',
-    'Snaku',
-    'Kaa',
-    'Madame Snake',
-    'Sir Hiss',
-    'Loki',
-    'Steelix',
-    'Gyarados',
-    'Seviper',
-    'Ekanes',
-    'Arbok',
-    'Snivy',
-    'Servine',
-    'Serperior',
-    'Mojo',
-    'Moss',
-    'Nigel',
-    'Tootsie',
-    'Sammy',
-    'Ziggy',
-    'Asmodeus',
-    'Attila',
-    'Basil',
-    'Diablo',
-    'Eden',
-    'Eve',
-    'Heaven',
-    'Hydra',
-    'Indiana',
-    'Jafaar',
-    'Kaa',
-    'Medusa',
-    'Naga',
-    'Severus',
-    'Slytherin',
-    'Snape',
-    'Raven',
-    'Slider',
-    'Slinky',
-    'Stripes',
-];
-
-export const DUCK_NAMES: ReadonlyArray<string> = [
-    'Quacky',
-    'Floaty',
-    'Duck',
-    'Molly',
-    'Sunshine',
-    'Buddy',
-    'Chirpy',
-    'Oscar',
-    'Lucy',
-    'Bailey',
-    'Beaky',
-    'Jemima',
-    'Peaches',
-    'Quackers',
-    'Jelly Beans',
-    'Donald',
-    'Chady',
-    'Waddles',
-    'Bill',
-    'Bubbles',
-    'James Pond',
-    'Moby Duck',
-    'Quack Sparrow',
-    'Peanut',
-    'Psyduck',
-    'Mr Quack',
-    'Louie',
-    'Golduck',
-    'Daisy',
-    'Pickles',
-    'Ducky Duck',
-    'Mrs Fluffs',
-    'Squeek',
-    'Ace',
-    'Rubberduck',
-    'Mrs Beak',
-    'April',
-    'Tutu',
-    'Billy the duck',
-    'Ducky',
-    'Neco',
-    'Dodo',
-    'Colonel',
-    'Franklin',
-    'Emmett',
-    'Bubba',
-    'Dillard',
-    'Duncan',
-    'Pogo',
-    'Uno',
-    'Peanut',
-    'Nero',
-    'Mowgli',
-    'Eggspresso',
-    'Webster',
-    'Quacker Jack',
-    'Plucker',
-    'Meeko',
-];
-
-export const ZAPPY_NAMES: ReadonlyArray<string> = [
-    'Zappy',
-    'Zippy',
-    'Zappy Jr.',
-    'Zoppy',
-    'Zuppy',
-    'Zeppy',
-    'Big Z',
-    'Little z',
-    'The Flash',
-    'Thor',
-    'Electric Bolt',
-    'Azula',
-    'Lightning Bolt',
-    'Power',
-    'Sonic',
-    'Speedy',
-    'Rush',
-];
-
-export const ROCKY_NAMES: ReadonlyArray<string> = [
-    'Rocky',
-    'The Rock',
-    'Quartzy',
-    'Rocky I',
-    'Rocky II',
-    'Rocky III',
-    'Pebbles Sr.',
-    'Big Granite',
-    'Boulder',
-    'Rockefeller',
-    'Pebble',
-    'Rocksanne',
-    'Rockstar',
-    'Onix',
-    'Rock and Roll',
-    'Dolomite',
-    'Granite',
-    'Miss Marble',
-    'Rock On',
-    'Amberstone',
-    'Rock With Me',
-    'Rock On It',
-    'Rock Out',
-];
-
-export const COCKATIEL_NAMES: ReadonlyArray<string> = [
-    'Cocktail',
-    'Pipsqueak',
-    'Sir Chirps a Lot',
-    'Nibbles',
-    'Lord of the Wings',
-    'Girl Nest Door',
-    'Wingman',
-    'Meryl Cheep',
-    'Jack Sparrow',
-    'Godfeather',
-    'Mickey',
-    'Baquack Obama',
-    'Dame Judi Finch',
-    'Kanye Nest',
-    'Speck',
-    'Cheecky',
-    'Arthur',
-    'Paco',
-    'Bobo',
-    'Walt',
-    'Happy',
-    'Junior',
-    'Coco',
-    'Yoyo',
-    'Milo',
-    'Skipper',
-    'Scarlet',
-    'Diva',
-    'Ursula',
-    'Donna',
-    'Lola',
-    'Kiko',
-    'Luna',
-];
 
 export function randomName(type: PetType): string {
     const collection: ReadonlyArray<string> =
@@ -548,6 +19,7 @@ export function randomName(type: PetType): string {
                 [PetType.dog]: DOG_NAMES,
                 [PetType.crab]: CRAB_NAMES,
                 [PetType.clippy]: CLIPPY_NAMES,
+                [PetType.mod]: MOD_NAMES,
                 [PetType.totoro]: TOTORO_NAMES,
                 [PetType.snake]: SNAKE_NAMES,
                 [PetType.rubberduck]: DUCK_NAMES,
