+++ src/common/types.ts
@@ -4,6 +4,7 @@ export const enum PetColor {
     green = 'green',
     yellow = 'yellow',
     gray = 'gray',
+    purple = 'purple',
     red = 'red',
     white = 'white',
     null = 'null',
@@ -15,6 +16,7 @@ export const enum PetType {
     cockatiel = 'cockatiel',
     crab = 'crab',
     dog = 'dog',
+    mod = 'mod',
     rocky = 'rocky',
     rubberduck = 'rubber-duck',
     snake = 'snake',
@@ -72,6 +74,7 @@ export const ALL_PETS = [
     PetType.cockatiel,
     PetType.crab,
     PetType.dog,
+    PetType.mod,
     PetType.rocky,
     PetType.rubberduck,
     PetType.snake,
@@ -84,6 +87,7 @@ export const ALL_COLORS = [
     PetColor.green,
     PetColor.yellow,
     PetColor.gray,
+    PetColor.purple,
     PetColor.red,
     PetColor.white,
     PetColor.null,
