+++ src/panel/pets.ts
@@ -1,18 +1,16 @@
 import { PetColor, PetSize, PetSpeed, PetType } from '../common/types';
-import { ISequenceTree } from './sequences';
-import {
-    IState,
-    States,
-    resolveState,
-    HorizontalDirection,
-    ChaseState,
-    BallState,
-    FrameResult,
-    PetInstanceState,
-    isStateAboveGround,
-} from './states';
-
-export class InvalidStateException {}
+import { Cat } from './pets/cat';
+import { Clippy } from './pets/clippy';
+import { Cockatiel } from './pets/cockatiel';
+import { Crab } from './pets/crab';
+import { Dog } from './pets/dog';
+import { Mod } from './pets/mod';
+import { Rocky } from './pets/rocky';
+import { RubberDuck } from './pets/rubberduck';
+import { Snake } from './pets/snake';
+import { Totoro } from './pets/totoro';
+import { Zappy } from './pets/zappy';
+import { IPetType } from './states';
 
 export class PetElement {
     el: HTMLImageElement;
@@ -135,897 +133,6 @@ export class PetCollection implements IPetCollection {
     }
 }
 
-export interface IPetType {
-    nextFrame(): void;
-
-    // Special methods for actions
-    canSwipe: boolean;
-    canChase: boolean;
-    swipe(): void;
-    chase(ballState: BallState, canvas: HTMLCanvasElement): void;
-    speed: number;
-    isMoving: boolean;
-    hello: string;
-
-    // State API
-    getState(): PetInstanceState;
-    recoverState(state: PetInstanceState): void;
-    recoverFriend(friend: IPetType): void;
-
-    // Positioning
-    bottom: number;
-    left: number;
-    positionBottom(bottom: number): void;
-    positionLeft(left: number): void;
-    width: number;
-    floor: number;
-
-    // Friends API
-    name: string;
-    emoji: string;
-    hasFriend: boolean;
-    friend: IPetType | undefined;
-    makeFriendsWith(friend: IPetType): boolean;
-    isPlaying: boolean;
-
-    showSpeechBubble(message: string, duration: number): void;
-}
-
-function calculateSpriteWidth(size: PetSize): number {
-    if (size === PetSize.nano) {
-        return 30;
-    } else if (size === PetSize.medium) {
-        return 55;
-    } else if (size === PetSize.large) {
-        return 110;
-    } else {
-        return 30; // Shrug
-    }
-}
-
-abstract class BasePetType implements IPetType {
-    label: string = 'base';
-    static count: number = 0;
-    sequence: ISequenceTree = {
-        startingState: States.sitIdle,
-        sequenceStates: [],
-    };
-    currentState: IState;
-    currentStateEnum: States;
-    holdState: IState | undefined;
-    holdStateEnum: States | undefined;
-    private el: HTMLImageElement;
-    private collision: HTMLDivElement;
-    private speech: HTMLDivElement;
-    private _left: number;
-    private _bottom: number;
-    petRoot: string;
-    _floor: number;
-    _friend: IPetType | undefined;
-    private _name: string;
-    private _speed: number;
-    private _size: PetSize;
-
-    constructor(
-        spriteElement: HTMLImageElement,
-        collisionElement: HTMLDivElement,
-        speechElement: HTMLDivElement,
-        size: PetSize,
-        left: number,
-        bottom: number,
-        petRoot: string,
-        floor: number,
-        name: string,
-        speed: number,
-    ) {
-        this.el = spriteElement;
-        this.collision = collisionElement;
-        this.speech = speechElement;
-        this.petRoot = petRoot;
-        this._floor = floor;
-        this._left = left;
-        this._bottom = bottom;
-        this.initSprite(size, left, bottom);
-        this.currentStateEnum = this.sequence.startingState;
-        this.currentState = resolveState(this.currentStateEnum, this);
-
-        this._name = name;
-        this._size = size;
-        this._speed = this.randomizeSpeed(speed);
-
-        // Increment the static count of the Pet class that the constructor belongs to
-        (this.constructor as any).count += 1;
-    }
-
-    initSprite(petSize: PetSize, left: number, bottom: number) {
-        this.el.style.left = `${left}px`;
-        this.el.style.bottom = `${bottom}px`;
-        this.el.style.width = 'auto';
-        this.el.style.height = 'auto';
-        this.el.style.maxWidth = `${calculateSpriteWidth(petSize)}px`;
-        this.el.style.maxHeight = `${calculateSpriteWidth(petSize)}px`;
-        this.collision.style.left = `${left}px`;
-        this.collision.style.bottom = `${bottom}px`;
-        this.collision.style.width = `${calculateSpriteWidth(petSize)}px`;
-        this.collision.style.height = `${calculateSpriteWidth(petSize)}px`;
-        this.speech.style.left = `${left}px`;
-        this.speech.style.bottom = `${
-            bottom + calculateSpriteWidth(petSize)
-        }px`;
-        this.hideSpeechBubble();
-    }
-
-    get left(): number {
-        return this._left;
-    }
-
-    get bottom(): number {
-        return this._bottom;
-    }
-
-    private repositionAccompanyingElements() {
-        this.collision.style.left = `${this._left}px`;
-        this.collision.style.bottom = `${this._bottom}px`;
-        this.speech.style.left = `${this._left}px`;
-        this.speech.style.bottom = `${
-            this._bottom + calculateSpriteWidth(this._size)
-        }px`;
-    }
-
-    positionBottom(bottom: number): void {
-        this._bottom = bottom;
-        this.el.style.bottom = `${this._bottom}px`;
-        this.repositionAccompanyingElements();
-    }
-
-    positionLeft(left: number): void {
-        this._left = left;
-        this.el.style.left = `${this._left}px`;
-        this.repositionAccompanyingElements();
-    }
-
-    get width(): number {
-        return this.el.width;
-    }
-
-    get floor(): number {
-        return this._floor;
-    }
-
-    get hello(): string {
-        // return the sound of the name of the animal
-        return ` says hello 👋!`;
-    }
-
-    getState(): PetInstanceState {
-        return { currentStateEnum: this.currentStateEnum };
-    }
-
-    get speed(): number {
-        return this._speed;
-    }
-
-    randomizeSpeed(speed: number): number {
-        const min = speed * 0.7;
-        const max = speed * 1.3;
-        const newSpeed = Math.random() * (max - min) + min;
-        return newSpeed;
-    }
-
-    get isMoving(): boolean {
-        return this._speed !== PetSpeed.still;
-    }
-
-    recoverFriend(friend: IPetType) {
-        // Recover friends..
-        this._friend = friend;
-    }
-
-    recoverState(state: PetInstanceState) {
-        // TODO : Resolve a bug where if it was swiping before, it would fail
-        // because holdState is no longer valid.
-        this.currentStateEnum = state.currentStateEnum ?? States.sitIdle;
-        this.currentState = resolveState(this.currentStateEnum, this);
-
-        if (!isStateAboveGround(this.currentStateEnum)) {
-            // Reset the bottom of the sprite to the floor as the theme
-            // has likely changed.
-            this.positionBottom(this.floor);
-        }
-    }
-
-    get canSwipe() {
-        return !isStateAboveGround(this.currentStateEnum);
-    }
-
-    get canChase() {
-        return (
-            !isStateAboveGround(this.currentStateEnum) &&
-            this.currentStateEnum !== States.chase &&
-            this.isMoving
-        );
-    }
-
-    showSpeechBubble(message: string, duration: number = 3000) {
-        this.speech.innerHTML = message;
-        this.speech.style.display = 'block';
-        setTimeout(() => {
-            this.hideSpeechBubble();
-        }, duration);
-    }
-
-    hideSpeechBubble() {
-        this.speech.style.display = 'none';
-    }
-
-    swipe() {
-        if (this.currentStateEnum === States.swipe) {
-            return;
-        }
-        this.holdState = this.currentState;
-        this.holdStateEnum = this.currentStateEnum;
-        this.currentStateEnum = States.swipe;
-        this.currentState = resolveState(this.currentStateEnum, this);
-        this.showSpeechBubble('👋');
-    }
-
-    chase(ballState: BallState, canvas: HTMLCanvasElement) {
-        this.currentStateEnum = States.chase;
-        this.currentState = new ChaseState(this, ballState, canvas);
-    }
-
-    faceLeft() {
-        this.el.style.transform = 'scaleX(-1)';
-    }
-
-    faceRight() {
-        this.el.style.transform = 'scaleX(1)';
-    }
-
-    setAnimation(face: string) {
-        if (this.el.src.endsWith(`_${face}_8fps.gif`)) {
-            return;
-        }
-        this.el.src = `${this.petRoot}_${face}_8fps.gif`;
-    }
-
-    chooseNextState(fromState: States): States {
-        // Work out next state
-        var possibleNextStates: States[] | undefined = undefined;
-        for (var i = 0; i < this.sequence.sequenceStates.length; i++) {
-            if (this.sequence.sequenceStates[i].state === fromState) {
-                possibleNextStates =
-                    this.sequence.sequenceStates[i].possibleNextStates;
-            }
-        }
-        if (!possibleNextStates) {
-            throw new InvalidStateException();
-        }
-        // randomly choose the next state
-        const idx = Math.floor(Math.random() * possibleNextStates.length);
-        return possibleNextStates[idx];
-    }
-
-    nextFrame() {
-        if (
-            this.currentState.horizontalDirection === HorizontalDirection.left
-        ) {
-            this.faceLeft();
-        } else if (
-            this.currentState.horizontalDirection === HorizontalDirection.right
-        ) {
-            this.faceRight();
-        }
-        this.setAnimation(this.currentState.spriteLabel);
-
-        // What's my buddy doing?
-        if (
-            this.hasFriend &&
-            this.currentStateEnum !== States.chaseFriend &&
-            this.isMoving
-        ) {
-            if (
-                this.friend?.isPlaying &&
-                !isStateAboveGround(this.currentStateEnum)
-            ) {
-                this.currentState = resolveState(States.chaseFriend, this);
-                this.currentStateEnum = States.chaseFriend;
-                return;
-            }
-        }
-
-        var frameResult = this.currentState.nextFrame();
-        if (frameResult === FrameResult.stateComplete) {
-            // If recovering from swipe..
-            if (this.holdState && this.holdStateEnum) {
-                this.currentState = this.holdState;
-                this.currentStateEnum = this.holdStateEnum;
-                this.holdState = undefined;
-                this.holdStateEnum = undefined;
-                return;
-            }
-
-            var nextState = this.chooseNextState(this.currentStateEnum);
-            this.currentState = resolveState(nextState, this);
-            this.currentStateEnum = nextState;
-        } else if (frameResult === FrameResult.stateCancel) {
-            if (this.currentStateEnum === States.chase) {
-                var nextState = this.chooseNextState(States.idleWithBall);
-                this.currentState = resolveState(nextState, this);
-                this.currentStateEnum = nextState;
-            } else if (this.currentStateEnum === States.chaseFriend) {
-                var nextState = this.chooseNextState(States.idleWithBall);
-                this.currentState = resolveState(nextState, this);
-                this.currentStateEnum = nextState;
-            }
-        }
-    }
-
-    get hasFriend(): boolean {
-        return this._friend !== undefined;
-    }
-
-    get friend(): IPetType | undefined {
-        return this._friend;
-    }
-
-    get name(): string {
-        return this._name;
-    }
-
-    makeFriendsWith(friend: IPetType): boolean {
-        this._friend = friend;
-        console.log(this.name, ": I'm now friends ❤️ with ", friend.name);
-        return true;
-    }
-
-    get isPlaying(): boolean {
-        return (
-            this.isMoving &&
-            (this.currentStateEnum === States.runRight ||
-                this.currentStateEnum === States.runLeft)
-        );
-    }
-
-    get emoji(): string {
-        return '🐶';
-    }
-}
-
-export class Totoro extends BasePetType {
-    label = 'totoro';
-    sequence = {
-        startingState: States.sitIdle,
-        sequenceStates: [
-            {
-                state: States.sitIdle,
-                possibleNextStates: [States.walkRight, States.lie],
-            },
-            {
-                state: States.lie,
-                possibleNextStates: [States.walkRight, States.walkLeft],
-            },
-            {
-                state: States.walkRight,
-                possibleNextStates: [States.walkLeft, States.sitIdle],
-            },
-            {
-                state: States.walkLeft,
-                possibleNextStates: [
-                    States.sitIdle,
-                    States.climbWallLeft,
-                    States.sitIdle,
-                ],
-            },
-            {
-                state: States.climbWallLeft,
-                possibleNextStates: [States.wallHangLeft],
-            },
-            {
-                state: States.wallHangLeft,
-                possibleNextStates: [States.jumpDownLeft],
-            },
-            {
-                state: States.jumpDownLeft,
-                possibleNextStates: [States.land],
-            },
-            {
-                state: States.land,
-                possibleNextStates: [
-                    States.sitIdle,
-                    States.walkRight,
-                    States.lie,
-                ],
-            },
-            {
-                state: States.chase,
-                possibleNextStates: [States.idleWithBall],
-            },
-            {
-                state: States.idleWithBall,
-                possibleNextStates: [States.walkRight, States.walkLeft],
-            },
-        ],
-    };
-    get emoji(): string {
-        return '🐾';
-    }
-    get hello(): string {
-        return `Try Laughing. Then Whatever Scares You Will Go Away. 🎭`;
-    }
-}
-export class Cat extends BasePetType {
-    label = 'cat';
-    sequence = {
-        startingState: States.sitIdle,
-        sequenceStates: [
-            {
-                state: States.sitIdle,
-                possibleNextStates: [States.walkRight, States.runRight],
-            },
-            {
-                state: States.walkRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.runRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.walkLeft,
-                possibleNextStates: [
-                    States.sitIdle,
-                    States.climbWallLeft,
-                    States.walkRight,
-                    States.runRight,
-                ],
-            },
-            {
-                state: States.runLeft,
-                possibleNextStates: [
-                    States.sitIdle,
-                    States.climbWallLeft,
-                    States.walkRight,
-                    States.runRight,
-                ],
-            },
-            {
-                state: States.climbWallLeft,
-                possibleNextStates: [States.wallHangLeft],
-            },
-            {
-                state: States.wallHangLeft,
-                possibleNextStates: [States.jumpDownLeft],
-            },
-            {
-                state: States.jumpDownLeft,
-                possibleNextStates: [States.land],
-            },
-            {
-                state: States.land,
-                possibleNextStates: [
-                    States.sitIdle,
-                    States.walkRight,
-                    States.runRight,
-                ],
-            },
-            {
-                state: States.chase,
-                possibleNextStates: [States.idleWithBall],
-            },
-            {
-                state: States.idleWithBall,
-                possibleNextStates: [
-                    States.walkRight,
-                    States.walkLeft,
-                    States.runLeft,
-                    States.runRight,
-                ],
-            },
-        ],
-    };
-    get emoji(): string {
-        return '🐱';
-    }
-    get hello(): string {
-        return `brrr... Meow!`;
-    }
-}
-
-export class Dog extends BasePetType {
-    label = 'dog';
-    sequence = {
-        startingState: States.sitIdle,
-        sequenceStates: [
-            {
-                state: States.sitIdle,
-                possibleNextStates: [
-                    States.walkRight,
-                    States.runRight,
-                    States.lie,
-                ],
-            },
-            {
-                state: States.lie,
-                possibleNextStates: [States.walkRight, States.runRight],
-            },
-            {
-                state: States.walkRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.runRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.walkLeft,
-                possibleNextStates: [
-                    States.sitIdle,
-                    States.lie,
-                    States.walkRight,
-                    States.runRight,
-                ],
-            },
-            {
-                state: States.runLeft,
-                possibleNextStates: [
-                    States.sitIdle,
-                    States.lie,
-                    States.walkRight,
-                    States.runRight,
-                ],
-            },
-            {
-                state: States.chase,
-                possibleNextStates: [States.idleWithBall],
-            },
-            {
-                state: States.idleWithBall,
-                possibleNextStates: [
-                    States.walkRight,
-                    States.walkLeft,
-                    States.runLeft,
-                    States.runRight,
-                ],
-            },
-        ],
-    };
-    get emoji(): string {
-        return '🐶';
-    }
-    get hello(): string {
-        return ` Every dog has its day - and today is woof day! Today I just want to bark. Take me on a walk`;
-    }
-}
-
-export class Snake extends BasePetType {
-    label = 'snake';
-    sequence = {
-        startingState: States.sitIdle,
-        sequenceStates: [
-            {
-                state: States.sitIdle,
-                possibleNextStates: [States.walkRight, States.runRight],
-            },
-            {
-                state: States.walkRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.runRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.walkLeft,
-                possibleNextStates: [
-                    States.sitIdle,
-                    States.walkRight,
-                    States.runRight,
-                ],
-            },
-            {
-                state: States.runLeft,
-                possibleNextStates: [
-                    States.sitIdle,
-                    States.walkRight,
-                    States.runRight,
-                ],
-            },
-            {
-                state: States.chase,
-                possibleNextStates: [States.idleWithBall],
-            },
-            {
-                state: States.idleWithBall,
-                possibleNextStates: [
-                    States.walkRight,
-                    States.walkLeft,
-                    States.runLeft,
-                    States.runRight,
-                ],
-            },
-        ],
-    };
-    get emoji(): string {
-        return '🐍';
-    }
-    get hello(): string {
-        return `Sss... Oh. Oh my gosh! I'm a snake!`;
-    }
-}
-
-export class Clippy extends BasePetType {
-    label = 'clippy';
-    sequence = {
-        startingState: States.sitIdle,
-        sequenceStates: [
-            {
-                state: States.sitIdle,
-                possibleNextStates: [States.walkRight, States.runRight],
-            },
-            {
-                state: States.walkRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.runRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.walkLeft,
-                possibleNextStates: [States.sitIdle],
-            },
-            {
-                state: States.runLeft,
-                possibleNextStates: [States.sitIdle],
-            },
-            {
-                state: States.chase,
-                possibleNextStates: [States.idleWithBall],
-            },
-            {
-                state: States.idleWithBall,
-                possibleNextStates: [
-                    States.walkRight,
-                    States.walkLeft,
-                    States.runLeft,
-                    States.runRight,
-                ],
-            },
-        ],
-    };
-    get emoji(): string {
-        return '📎';
-    }
-    get hello(): string {
-        return ` Hi, I'm Clippy, would you like some assistance today? 👋!`;
-    }
-}
-
-export class RubberDuck extends BasePetType {
-    label = 'rubber-duck';
-    sequence = {
-        startingState: States.sitIdle,
-        sequenceStates: [
-            {
-                state: States.sitIdle,
-                possibleNextStates: [States.walkRight, States.runRight],
-            },
-            {
-                state: States.walkRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.runRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.walkLeft,
-                possibleNextStates: [States.sitIdle],
-            },
-            {
-                state: States.runLeft,
-                possibleNextStates: [States.sitIdle],
-            },
-            {
-                state: States.chase,
-                possibleNextStates: [States.idleWithBall],
-            },
-            {
-                state: States.idleWithBall,
-                possibleNextStates: [
-                    States.walkRight,
-                    States.walkLeft,
-                    States.runLeft,
-                    States.runRight,
-                ],
-            },
-        ],
-    };
-    get emoji(): string {
-        return '🐥';
-    }
-    get hello(): string {
-        return ` Hi, I love to quack around 👋!`;
-    }
-}
-
-export class Cockatiel extends BasePetType {
-    label = 'cockatiel';
-    sequence = {
-        startingState: States.sitIdle,
-        sequenceStates: [
-            {
-                state: States.sitIdle,
-                possibleNextStates: [States.walkRight, States.runRight],
-            },
-            {
-                state: States.walkRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.runRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.walkLeft,
-                possibleNextStates: [States.sitIdle],
-            },
-            {
-                state: States.runLeft,
-                possibleNextStates: [States.sitIdle],
-            },
-            {
-                state: States.chase,
-                possibleNextStates: [States.idleWithBall],
-            },
-            {
-                state: States.idleWithBall,
-                possibleNextStates: [
-                    States.walkRight,
-                    States.walkLeft,
-                    States.runLeft,
-                    States.runRight,
-                ],
-            },
-        ],
-    };
-    get emoji(): string {
-        return '🦜';
-    }
-    get hello(): string {
-        // TODO: #191 Add a custom message for cockatiel
-        return ` Hello, I'm a good bird 👋!`;
-    }
-}
-
-export class Crab extends BasePetType {
-    label = 'crab';
-    sequence = {
-        startingState: States.sitIdle,
-        sequenceStates: [
-            {
-                state: States.sitIdle,
-                possibleNextStates: [States.walkRight, States.runRight],
-            },
-            {
-                state: States.walkRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.runRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.walkLeft,
-                possibleNextStates: [States.sitIdle],
-            },
-            {
-                state: States.runLeft,
-                possibleNextStates: [States.sitIdle],
-            },
-            {
-                state: States.chase,
-                possibleNextStates: [States.idleWithBall],
-            },
-            {
-                state: States.idleWithBall,
-                possibleNextStates: [
-                    States.walkRight,
-                    States.walkLeft,
-                    States.runLeft,
-                    States.runRight,
-                ],
-            },
-        ],
-    };
-    get emoji(): string {
-        return '🦀';
-    }
-    get hello(): string {
-        return ` Hi, I'm Crabsolutely Clawsome Crab 👋!`;
-    }
-}
-
-export class Zappy extends BasePetType {
-    label = 'zappy';
-    sequence = {
-        startingState: States.sitIdle,
-        sequenceStates: [
-            {
-                state: States.sitIdle,
-                possibleNextStates: [States.walkRight, States.runRight],
-            },
-            {
-                state: States.walkRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.runRight,
-                possibleNextStates: [States.walkLeft, States.runLeft],
-            },
-            {
-                state: States.walkLeft,
-                possibleNextStates: [States.sitIdle],
-            },
-            {
-                state: States.runLeft,
-                possibleNextStates: [States.sitIdle],
-            },
-            {
-                state: States.chase,
-                possibleNextStates: [States.idleWithBall],
-            },
-            {
-                state: States.idleWithBall,
-                possibleNextStates: [
-                    States.walkRight,
-                    States.walkLeft,
-                    States.runLeft,
-                    States.runRight,
-                ],
-            },
-        ],
-    };
-    get emoji(): string {
-        return '⚡';
-    }
-    get hello(): string {
-        // TODO: #193 Add a custom message for zappy
-        return ` Hello this is Zappy! Do I look familiar?? I am the mascot for Azure Functions😉`;
-    }
-}
-
-export class Rocky extends BasePetType {
-    label = 'rocky';
-    sequence = {
-        startingState: States.sitIdle,
-        sequenceStates: [
-            {
-                state: States.sitIdle,
-                possibleNextStates: [States.walkRight, States.runRight],
-            },
-            {
-                state: States.walkRight,
-                possibleNextStates: [States.sitIdle, States.runRight],
-            },
-            {
-                state: States.runRight,
-                possibleNextStates: [States.sitIdle, States.walkRight],
-            },
-        ],
-    };
-    get emoji(): string {
-        return '💎';
-    }
-    get canChase(): boolean {
-        return false;
-    }
-    get hello(): string {
-        return ` 👋 I'm rock! I always Rock`;
-    }
-}
-
 export class InvalidPetException {
     message?: string;
 
@@ -1071,6 +178,8 @@ export function createPet(
             return new Crab(...standardPetArguments, PetSpeed.slow);
         case PetType.clippy:
             return new Clippy(...standardPetArguments, PetSpeed.slow);
+        case PetType.mod:
+            return new Mod(...standardPetArguments, PetSpeed.normal);
         case PetType.totoro:
             return new Totoro(...standardPetArguments, PetSpeed.normal);
         case PetType.snake:
