+++ src/panel/pets/cockatiel.ts
@@ -1,0 +1,87 @@
+import { BasePetType } from '../basepettype';
+import { States } from '../states';
+
+export class Cockatiel extends BasePetType {
+    label = 'cockatiel';
+    sequence = {
+        startingState: States.sitIdle,
+        sequenceStates: [
+            {
+                state: States.sitIdle,
+                possibleNextStates: [States.walkRight, States.runRight],
+            },
+            {
+                state: States.walkRight,
+                possibleNextStates: [States.walkLeft, States.runLeft],
+            },
+            {
+                state: States.runRight,
+                possibleNextStates: [States.walkLeft, States.runLeft],
+            },
+            {
+                state: States.walkLeft,
+                possibleNextStates: [States.sitIdle],
+            },
+            {
+                state: States.runLeft,
+                possibleNextStates: [States.sitIdle],
+            },
+            {
+                state: States.chase,
+                possibleNextStates: [States.idleWithBall],
+            },
+            {
+                state: States.idleWithBall,
+                possibleNextStates: [
+                    States.walkRight,
+                    States.walkLeft,
+                    States.runLeft,
+                    States.runRight,
+                ],
+            },
+        ],
+    };
+    get emoji(): string {
+        return '🦜';
+    }
+    get hello(): string {
+        // TODO: #191 Add a custom message for cockatiel
+        return ` Hello, I'm a good bird 👋!`;
+    }
+}
+
+export const COCKATIEL_NAMES: ReadonlyArray<string> = [
+    'Cocktail',
+    'Pipsqueak',
+    'Sir Chirps a Lot',
+    'Nibbles',
+    'Lord of the Wings',
+    'Girl Nest Door',
+    'Wingman',
+    'Meryl Cheep',
+    'Jack Sparrow',
+    'Godfeather',
+    'Mickey',
+    'Baquack Obama',
+    'Dame Judi Finch',
+    'Kanye Nest',
+    'Speck',
+    'Cheecky',
+    'Arthur',
+    'Paco',
+    'Bobo',
+    'Walt',
+    'Happy',
+    'Junior',
+    'Coco',
+    'Yoyo',
+    'Milo',
+    'Skipper',
+    'Scarlet',
+    'Diva',
+    'Ursula',
+    'Donna',
+    'Lola',
+    'Kiko',
+    'Luna',
+];
