+++ src/panel/pets/dog.ts
@@ -1,0 +1,208 @@
+import { BasePetType } from '../basepettype';
+import { States } from '../states';
+
+export class Dog extends BasePetType {
+    label = 'dog';
+    sequence = {
+        startingState: States.sitIdle,
+        sequenceStates: [
+            {
+                state: States.sitIdle,
+                possibleNextStates: [
+                    States.walkRight,
+                    States.runRight,
+                    States.lie,
+                ],
+            },
+            {
+                state: States.lie,
+                possibleNextStates: [States.walkRight, States.runRight],
+            },
+            {
+                state: States.walkRight,
+                possibleNextStates: [States.walkLeft, States.runLeft],
+            },
+            {
+                state: States.runRight,
+                possibleNextStates: [States.walkLeft, States.runLeft],
+            },
+            {
+                state: States.walkLeft,
+                possibleNextStates: [
+                    States.sitIdle,
+                    States.lie,
+                    States.walkRight,
+                    States.runRight,
+                ],
+            },
+            {
+                state: States.runLeft,
+                possibleNextStates: [
+                    States.sitIdle,
+                    States.lie,
+                    States.walkRight,
+                    States.runRight,
+                ],
+            },
+            {
+                state: States.chase,
+                possibleNextStates: [States.idleWithBall],
+            },
+            {
+                state: States.idleWithBall,
+                possibleNextStates: [
+                    States.walkRight,
+                    States.walkLeft,
+                    States.runLeft,
+                    States.runRight,
+                ],
+            },
+        ],
+    };
+    get emoji(): string {
+        return '🐶';
+    }
+    get hello(): string {
+        return ` Every dog has its day - and today is woof day! Today I just want to bark. Take me on a walk`;
+    }
+}
+
+export const DOG_NAMES: ReadonlyArray<string> = [
+    'Bella',
+    'Charlie',
+    'Max',
+    'Molly',
+    'Coco',
+    'Buddy',
+    'Ruby',
+    'Oscar',
+    'Lucy',
+    'Bailey',
+    'Milo',
+    'Daisy',
+    'Archie',
+    'Ollie',
+    'Rosie',
+    'Lola',
+    'Frankie',
+    'Toby',
+    'Roxy',
+    'Poppy',
+    'Luna',
+    'Jack',
+    'Millie',
+    'Teddy',
+    'Harry',
+    'Cooper',
+    'Bear',
+    'Rocky',
+    'Alfie',
+    'Hugo',
+    'Bonnie',
+    'Pepper',
+    'Lily',
+    'Leo',
+    'Maggie',
+    'George',
+    'Mia',
+    'Marley',
+    'Harley',
+    'Chloe',
+    'Lulu',
+    'Jasper',
+    'Billy',
+    'Nala',
+    'Monty',
+    'Ziggy',
+    'Winston',
+    'Zeus',
+    'Zoe',
+    'Stella',
+    'Sasha',
+    'Rusty',
+    'Gus',
+    'Baxter',
+    'Dexter',
+    'Diesel',
+    'Willow',
+    'Barney',
+    'Bruno',
+    'Penny',
+    'Honey',
+    'Milly',
+    'Murphy',
+    'Holly',
+    'Benji',
+    'Henry',
+    'Lilly',
+    'Pippa',
+    'Shadow',
+    'Sam',
+    'Buster',
+    'Lucky',
+    'Ellie',
+    'Duke',
+    'Jessie',
+    'Cookie',
+    'Harvey',
+    'Bruce',
+    'Jax',
+    'Rex',
+    'Louie',
+    'Bentley',
+    'Jet',
+    'Banjo',
+    'Beau',
+    'Ella',
+    'Ralph',
+    'Loki',
+    'Lexi',
+    'Chester',
+    'Sophie',
+    'Billie',
+    'Louis',
+    'Charlie',
+    'Cleo',
+    'Spot',
+    'Harry',
+    'Bolt',
+    'Ein',
+    'Maddy',
+    'Ghost',
+    'Midnight',
+    'Pumpkin',
+    'Shadow',
+    'Sparky',
+    'Linus',
+    'Cody',
+    'Slinky',
+    'Toto',
+    'Balto',
+    'Golfo',
+    'Pongo',
+    'Beethoven',
+    'Hachiko',
+    'Scooby',
+    'Clifford',
+    'Astro',
+    'Goofy',
+    'Chip',
+    'Einstein',
+    'Fang',
+    'Truman',
+    'Uggie',
+    'Bingo',
+    'Blue',
+    'Cometa',
+    'Krypto',
+    'Huesos',
+    'Odie',
+    'Snoopy',
+    'Aisha',
+    'Moly',
+    'Chiquita',
+    'Chavela',
+    'Tramp',
+    'Lady',
+    'Puddles',
+];
