+++ src/panel/pets/rubberduck.ts
@@ -1,0 +1,111 @@
+import { BasePetType } from '../basepettype';
+import { States } from '../states';
+
+export class RubberDuck extends BasePetType {
+    label = 'rubber-duck';
+    sequence = {
+        startingState: States.sitIdle,
+        sequenceStates: [
+            {
+                state: States.sitIdle,
+                possibleNextStates: [States.walkRight, States.runRight],
+            },
+            {
+                state: States.walkRight,
+                possibleNextStates: [States.walkLeft, States.runLeft],
+            },
+            {
+                state: States.runRight,
+                possibleNextStates: [States.walkLeft, States.runLeft],
+            },
+            {
+                state: States.walkLeft,
+                possibleNextStates: [States.sitIdle],
+            },
+            {
+                state: States.runLeft,
+                possibleNextStates: [States.sitIdle],
+            },
+            {
+                state: States.chase,
+                possibleNextStates: [States.idleWithBall],
+            },
+            {
+                state: States.idleWithBall,
+                possibleNextStates: [
+                    States.walkRight,
+                    States.walkLeft,
+                    States.runLeft,
+                    States.runRight,
+                ],
+            },
+        ],
+    };
+    get emoji(): string {
+        return '🐥';
+    }
+    get hello(): string {
+        return ` Hi, I love to quack around 👋!`;
+    }
+}
+
+export const DUCK_NAMES: ReadonlyArray<string> = [
+    'Quacky',
+    'Floaty',
+    'Duck',
+    'Molly',
+    'Sunshine',
+    'Buddy',
+    'Chirpy',
+    'Oscar',
+    'Lucy',
+    'Bailey',
+    'Beaky',
+    'Jemima',
+    'Peaches',
+    'Quackers',
+    'Jelly Beans',
+    'Donald',
+    'Chady',
+    'Waddles',
+    'Bill',
+    'Bubbles',
+    'James Pond',
+    'Moby Duck',
+    'Quack Sparrow',
+    'Peanut',
+    'Psyduck',
+    'Mr Quack',
+    'Louie',
+    'Golduck',
+    'Daisy',
+    'Pickles',
+    'Ducky Duck',
+    'Mrs Fluffs',
+    'Squeek',
+    'Ace',
+    'Rubberduck',
+    'Mrs Beak',
+    'April',
+    'Tutu',
+    'Billy the duck',
+    'Ducky',
+    'Neco',
+    'Dodo',
+    'Colonel',
+    'Franklin',
+    'Emmett',
+    'Bubba',
+    'Dillard',
+    'Duncan',
+    'Pogo',
+    'Uno',
+    'Peanut',
+    'Nero',
+    'Mowgli',
+    'Eggspresso',
+    'Webster',
+    'Quacker Jack',
+    'Plucker',
+    'Meeko',
+];
