+++ src/panel/pets/snake.ts
@@ -1,0 +1,113 @@
+import { BasePetType } from '../basepettype';
+import { States } from '../states';
+
+export class Snake extends BasePetType {
+    label = 'snake';
+    sequence = {
+        startingState: States.sitIdle,
+        sequenceStates: [
+            {
+                state: States.sitIdle,
+                possibleNextStates: [States.walkRight, States.runRight],
+            },
+            {
+                state: States.walkRight,
+                possibleNextStates: [States.walkLeft, States.runLeft],
+            },
+            {
+                state: States.runRight,
+                possibleNextStates: [States.walkLeft, States.runLeft],
+            },
+            {
+                state: States.walkLeft,
+                possibleNextStates: [
+                    States.sitIdle,
+                    States.walkRight,
+                    States.runRight,
+                ],
+            },
+            {
+                state: States.runLeft,
+                possibleNextStates: [
+                    States.sitIdle,
+                    States.walkRight,
+                    States.runRight,
+                ],
+            },
+            {
+                state: States.chase,
+                possibleNextStates: [States.idleWithBall],
+            },
+            {
+                state: States.idleWithBall,
+                possibleNextStates: [
+                    States.walkRight,
+                    States.walkLeft,
+                    States.runLeft,
+                    States.runRight,
+                ],
+            },
+        ],
+    };
+    get emoji(): string {
+        return '🐍';
+    }
+    get hello(): string {
+        return `Sss... Oh. Oh my gosh! I'm a snake!`;
+    }
+}
+
+export const SNAKE_NAMES: ReadonlyArray<string> = [
+    'Sneaky',
+    'Mr Slippery',
+    'Hissy Elliott',
+    'Molly',
+    'Coco',
+    'Buddy',
+    'Ruby',
+    'Bailey',
+    'Max',
+    'Seb',
+    'Kaa',
+    'Mr Hiss',
+    'Miss Hiss',
+    'Snaku',
+    'Kaa',
+    'Madame Snake',
+    'Sir Hiss',
+    'Loki',
+    'Steelix',
+    'Gyarados',
+    'Seviper',
+    'Ekanes',
+    'Arbok',
+    'Snivy',
+    'Servine',
+    'Serperior',
+    'Mojo',
+    'Moss',
+    'Nigel',
+    'Tootsie',
+    'Sammy',
+    'Ziggy',
+    'Asmodeus',
+    'Attila',
+    'Basil',
+    'Diablo',
+    'Eden',
+    'Eve',
+    'Heaven',
+    'Hydra',
+    'Indiana',
+    'Jafaar',
+    'Kaa',
+    'Medusa',
+    'Naga',
+    'Severus',
+    'Slytherin',
+    'Snape',
+    'Raven',
+    'Slider',
+    'Slinky',
+    'Stripes',
+];
