+++ src/panel/pets/zappy.ts
@@ -1,0 +1,71 @@
+import { BasePetType } from '../basepettype';
+import { States } from '../states';
+
+export class Zappy extends BasePetType {
+    label = 'zappy';
+    sequence = {
+        startingState: States.sitIdle,
+        sequenceStates: [
+            {
+                state: States.sitIdle,
+                possibleNextStates: [States.walkRight, States.runRight],
+            },
+            {
+                state: States.walkRight,
+                possibleNextStates: [States.walkLeft, States.runLeft],
+            },
+            {
+                state: States.runRight,
+                possibleNextStates: [States.walkLeft, States.runLeft],
+            },
+            {
+                state: States.walkLeft,
+                possibleNextStates: [States.sitIdle],
+            },
+            {
+                state: States.runLeft,
+                possibleNextStates: [States.sitIdle],
+            },
+            {
+                state: States.chase,
+                possibleNextStates: [States.idleWithBall],
+            },
+            {
+                state: States.idleWithBall,
+                possibleNextStates: [
+                    States.walkRight,
+                    States.walkLeft,
+                    States.runLeft,
+                    States.runRight,
+                ],
+            },
+        ],
+    };
+    get emoji(): string {
+        return '⚡';
+    }
+    get hello(): string {
+        // TODO: #193 Add a custom message for zappy
+        return ` Hello this is Zappy! Do I look familiar?? I am the mascot for Azure Functions😉`;
+    }
+}
+
+export const ZAPPY_NAMES: ReadonlyArray<string> = [
+    'Zappy',
+    'Zippy',
+    'Zappy Jr.',
+    'Zoppy',
+    'Zuppy',
+    'Zeppy',
+    'Big Z',
+    'Little z',
+    'The Flash',
+    'Thor',
+    'Electric Bolt',
+    'Azula',
+    'Lightning Bolt',
+    'Power',
+    'Sonic',
+    'Speedy',
+    'Rush',
+];
